Style splash chatbot, refactor davinci code to clarify model name, add handler for turbo model overload
