+++ backend/app.js
@@ -18,7 +18,7 @@ const usersRouter = require("./routes/api/users");
 const csrfRouter = require("./routes/api/csrf");
 const favoritesRouter = require("./routes/api/favorites");
 const reviewsRouter = require("./routes/api/reviews");
-const aiRouter = require("./routes/api/generate");
+const aiRouterDavinci = require("./routes/api/generateDavinci");
 const aiRouterTurbo = require("./routes/api/generateTurbo");
 const recipesRouter = require("./routes/api/recipes");
 const searchRouter = require("./routes/api/search");
@@ -47,7 +47,7 @@ app.use(
 
 app.use("/api/users", usersRouter);
 app.use("/api/csrf", csrfRouter);
-app.use("/api/generate", aiRouter);
+app.use("/api/generateDavinci", aiRouterDavinci);
 app.use("/api/generateTurbo", aiRouterTurbo);
 app.use("/api/favorites", favoritesRouter);
 app.use("/api/reviews", reviewsRouter);
